Finished edit toggle
